pyproject.toml
[tool.poetry]
name = "repl-nix-lambda-calculus-polyglot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "repl_nix_lambda_calculus_polyglot"}]

[tool.poetry.dependencies]
python = "^3.12"
click = "^8.1.7"
jinja2 = "^3.1.4"
openai = "^1.37.1"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
